+++ internal/gojs/fs_test.go
@@ -33,7 +33,7 @@ func Test_writefs(t *testing.T) {
 	t.Parallel()
 	tmpDir := t.TempDir()
 
-	fs := writefs.New(tmpDir)
+	fs := writefs.DirFS(tmpDir)
 	// test expects to write under /tmp
 	require.NoError(t, os.Mkdir(path.Join(tmpDir, "tmp"), 0o700))
